requirements.txt
alibabacloud_oss_v2==1.1.2
APScheduler==3.11.0
beautifulsoup4>=4.13.5
seleniumbase==4.45.6
python-telegram-bot==22.3
PyYAML>=6.0.2
qrcode==8.2
Requests==2.32.5
telethon[cryptg]==1.40.0
urllib3==2.5.0
Pillow==11.2.1
aria2p==0.12.1
PySocks==1.7.1
